Revert "Merge branch 'master' of https://github.com/FitzerIRL/pxCore"
This reverts commit 0fe607bd54c8c61a887bfb106dcf4de6b7b8139f, reversing
changes made to 99063c3de876c20104fd32043d1d97045a7a8ac8.

+++ examples/pxScene2d/src/pxResource.cpp
@@ -151,7 +151,7 @@ void pxResource::raiseDownloadPriority()
   {
     rtLogWarn(">>>>>>>>>>>>>>>>>>>>>>>Inside pxResource::raiseDownloadPriority and download is in progress for %s\n",mUrl.cString());
     mListenersMutex.lock();
-    size_t lisenersSize = mListeners.size();
+    int lisenersSize = mListeners.size();
     mListenersMutex.unlock();
     if( lisenersSize == 0) 
       rtLogInfo("But size is 0, so no one cares!!!!!\n");
